refactor: centralize jwt secret and improve registration
